updated route structure
