+++ academy-ui/src/components/courses/View/Details/QuestionDetails.tsx
@@ -135,8 +135,8 @@ export default function QuestionDetails(props: QuestionDetailsProps) {
 
   const { editMode, cancel, showEdit, save } = useEditCourseDetails<UpdateTopicQuestionInput>(
     async (updates: UpdateTopicQuestionInput) =>
-      await fetch('/api/courses/update-topic-question', {
-        method: 'POST',
+      await fetch(`/api/courses/${updates.courseKey}/topics/${updates.topicKey}/question`, {
+        method: 'PUT',
         body: JSON.stringify({
           spaceId: space.id,
           questionInfo: updates,
@@ -146,8 +146,8 @@ export default function QuestionDetails(props: QuestionDetailsProps) {
 
   const { deleting, deleteItem } = useDeleteCourseItem<DeleteTopicQuestionInput>(
     async (updates: DeleteTopicQuestionInput) =>
-      await fetch('/api/courses/delete-topic-question', {
-        method: 'POST',
+      await fetch(`/api/courses/${updates.courseKey}/topics/${updates.topicKey}/question`, {
+        method: 'DELETE',
         body: JSON.stringify({
           spaceId: space.id,
           questionInfo: updates,
@@ -163,8 +163,8 @@ export default function QuestionDetails(props: QuestionDetailsProps) {
 
   const { movingUp, movingDown, moveItem } = useMoveCourseItem<MoveTopicQuestionInput>(
     async (updates: MoveTopicQuestionInput) =>
-      await fetch('/api/courses/move-topic-question', {
-        method: 'POST',
+      await fetch(`/api/courses/${updates.courseKey}/topics/${updates.topicKey}/question`, {
+        method: 'PATCH',
         body: JSON.stringify({
           spaceId: space.id,
           questionInfo: updates,
